fix: update navigation links
