+++ src/app/components/atoms/title/title.component.css
@@ -6,7 +6,7 @@ h2.title-section {
   width: 100%;
 }
 
-h3.sub-title {
+.sub-title {
   opacity: 0.8;
   font-size: 1.1rem;
   font-weight: 500;
