+++ src/app/components/legacy/bento-item-expertice/bento-item-expertice.component.ts
