+++ src/app/components/legacy/studies/studies.component.spec.ts
