+++ src/app/components/legacy/studies/studies.component.ts
