+++ src/app/pages/home/layouts/experiency/experiency.component.html
@@ -1,4 +1,4 @@
-<section id="experiency">
+<section id="career">
   <app-title
   className="text-center"
   titleID="experiencyTitle"
